+++ ChangeLog.md
@@ -21,6 +21,7 @@
 - Fix issue with missing known DicomTransferSyntax from static DicomTransferSyntax.Entries dictionary (#1644)
 - Improve robustness of DicomService when presented with HTTP requests. Bail early if the PDU type is not recognized (#1678)
 - Enhancement: Added IEquatable implementation and equality operators for DicomDataset class
+- Fix issue where a DICOM server could stop accepting incoming connections if MaxClientsAllowed is configured and one or more connections take longer than one minute to close (#1670)
 
 #### 5.1.1 (2023-05-29)